+++ src/components/Modal.js
@@ -1,0 +1,20 @@
+import React, {Component} from 'react';
+import * as Icon from 'react-feather';
+
+ const Modal = ({handleClose,show,children}) => {
+    const showHideClassName = show ? 'modal display-block':'model display-none';
+
+    return(
+        <div className={showHideClassName}>
+        <section className='modal-main'>
+            <Icon.X className="modal_button_close pointer" onClick={handleClose}/>
+        
+        {children}
+        
+        
+        </section>
+        </div>
+    )
+}
+
+export default Modal;
