+++ src/styles/DriverInfo.css
@@ -31,7 +31,7 @@ img{
 .modal_button{
     width:200px;
     height:30px;
-    margin-left: 500px;
+    margin-left: 10px;
     margin-top: 50px;
     margin-bottom: 10px;
 }
